\name{links}
\alias{links}
\docType{data}
\title{
Data: Links between Italian families in the Renaissance.
}
\description{
Data frame with the marriage and business links.
}
\usage{data("links")}
\format{
  A data frame with 36 links (rows) amongst 16 Italian families in the Renaissance.
  \describe{
    \item{\code{Albizzi}}{}
    \item{\code{Acciaiuoli}}{}
    \item{\code{Barbadori}}{}
    \item{\code{Bischeri}}{}
    \item{\code{Castellani}}{}
    \item{\code{Guadagni}}{}
    \item{\code{Lamberteschi}}{}
    \item{\code{Medici}}{}
    \item{\code{Pazzi}}{}
    \item{\code{Peruzzi}}{}
    \item{\code{Ridolfi}}{}
    \item{\code{Salviati}}{}
    \item{\code{Strozzi}}{}
    \item{\code{Tornabuoni}}{}
    \item{\code{Ginori}}{}
    \item{\code{Pucci}}{}
    \item{\code{link}}{Type of link: marriage or business}    
  }
}
\source{
PADGETT, J. F. Y C. K. ANSELL (1993): "Robust Action and the Rise of the Medici, 1400-1434", in American Journal of Sociology, 98, 1259-1319. (http://www.jstor.org/stable/2781822)
}

\examples{
data("links")
head(links)
}
\keyword{datasets}
